Add LongColumn and remove the need for data wrappers (#350)
* Remove the need for data wrappers

* Add LongColumn

+++ core/src/main/java/tech/tablesaw/columns/datetimes/PackedLocalDateTime.java
@@ -392,7 +392,7 @@ public class PackedLocalDateTime {
         return PackedLocalTime.getSecond(time(packedLocalDateTime));
     }
 
-    public static double getSecondOfDay(long packedLocalDateTime) {
+    public static int getSecondOfDay(long packedLocalDateTime) {
         return PackedLocalTime.getSecondOfDay(time(packedLocalDateTime));
     }
