+++ core/src/test/java/tech/tablesaw/io/csv/CsvReaderTest.java
@@ -48,8 +48,8 @@ import org.junit.rules.ExpectedException;
 import tech.tablesaw.api.ColumnType;
 import tech.tablesaw.api.DateColumn;
 import tech.tablesaw.api.DateTimeColumn;
-import tech.tablesaw.api.DoubleColumn;
 import tech.tablesaw.api.IntColumn;
+import tech.tablesaw.api.LongColumn;
 import tech.tablesaw.api.Table;
 
 /**
@@ -143,9 +143,8 @@ public class CsvReaderTest {
 
     @Test
     public void testMillis() {
-
         long[] times = {1530486314124L, 1530488214124L};
-        DoubleColumn d = DoubleColumn.create("times", times);
+        LongColumn d = LongColumn.create("times", times);
         DateTimeColumn column = d.asDateTimes(ZoneOffset.UTC);
         assertEquals(1530486314124L, column.get(0).toInstant(ZoneOffset.UTC).toEpochMilli());
     }
